feat: add clan wars safe and dangerous free-for-all portals.
