+++ data/minigame/clan_wars/clan_wars.areas.toml
@@ -4,7 +4,7 @@ y = [3679, 3682]
 tags = ["teleport"]
 
 [clan_wars_ffa]
-x = [2740, 2890]
+x = [2740, 3090]
 y = [5490, 5640]
 
 [clan_wars_ffa_safe_arena]
@@ -14,4 +14,13 @@ y = [5512, 5630]
 [clan_wars_ffa_safe_multi]
 x = [2756, 2878]
 y = [5571, 5630]
+tags = ["multi_combat"]
+
+[clan_wars_ffa_dangerous_arena]
+x = [2948, 3071]
+y = [5512, 5631]
+
+[clan_wars_ffa_dangerous_multi]
+x = [2948, 3071]
+y = [5571, 5631]
 tags = ["multi_combat"]
