+++ data/minigame/clan_wars/clan_wars.varbits.toml
@@ -1,0 +1,3 @@
+[clan_wars_ffa_portal]
+id = 5279
+format = "int"
